more updates to the build scripts.
git-svn-id: file:///svn/uhexen2/code/trunk@2973 eaf649e6-6c13-0410-b194-84221022858e

+++ hexenworld/build.sh
@@ -1,12 +1,17 @@
 #!/bin/sh
 
-EXE_EXT=
-BIN_FILES="Master/hwmaster Server/hwsv Client/hwcl Client/glhwcl"
-
-if env | grep -i windir > __tmp.tmp; then
-EXE_EXT=".exe";
+if [ "$1" = "strip" ]; then
+	exe_ext=
+	if env | grep -i windir > __tmp.tmp; then
+		exe_ext=".exe"
+	fi
+	rm -f __tmp.tmp
+	strip Master/hwmaster$exe_ext	\
+		Server/hwsv$exe_ext	\
+		Client/hwcl$exe_ext	\
+		Client/glhwcl$exe_ext
+	exit 0
 fi
-rm -f __tmp.tmp
 
 HOST_OS=`uname|sed -e s/_.*//|tr '[:upper:]' '[:lower:]'`
 
@@ -22,21 +27,11 @@ linux)
 	;;
 esac
 
-if [ "$1" = "strip" ]
-then
-	for i in ${BIN_FILES}
-	do
-	    strip ${i}${EXE_EXT}
-	done
-exit 0
-fi
-
-if [ "$1" = "clean" ]
-then
-$MAKE_CMD -s -C Client clean
-$MAKE_CMD -s -C Master clean
-$MAKE_CMD -s -C Server clean
-exit 0
+if [ "$1" = "clean" ]; then
+	$MAKE_CMD -s -C Client clean
+	$MAKE_CMD -s -C Master clean
+	$MAKE_CMD -s -C Server clean
+	exit 0
 fi
 
 echo "Building hexenworld server..."
